requirements.txt
Flask>=2.0
Pillow>=9.0
cryptography>=38.0
gunicorn
